Fix All 🚀🚀🚑

+++ LibraryMS/Repository/IBookRepository.cs
@@ -6,10 +6,10 @@ namespace LibraryMS.Repository
     public interface IBookRepository
     {
         Task<IEnumerable<Book>> GetAllBookAsync(CancellationToken cancellationToken);
-        Task<Book> GetBookByIdAsync(int id, CancellationToken cancellationToken);
+        Task<Book> GetBookByIdAsync(long id, CancellationToken cancellationToken);
         Task<Book> AddBookAsync(Book book, CancellationToken cancellationToken);
         Task<Book> UpdateBookAsync(Book book, CancellationToken cancellationToken);
-        Task<Book> DeleteBooktAsync(int id, CancellationToken cancellationToken);
+        Task<Book> DeleteBooktAsync(long id, CancellationToken cancellationToken);
         IEnumerable<SelectListItem> Dropdown();
     }
 }
